feat: shared ui 폴더 구분 및 app layer 파일 이동
