+++ src/shared/ui/card/card.ui.tsx
@@ -1,0 +1,9 @@
+import { forwardRef } from "react"
+
+interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string
+}
+
+export const Card = forwardRef<HTMLDivElement, CardProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+))
