+++ src/shared/ui/card/cardContent.ui.tsx
@@ -1,0 +1,9 @@
+import { forwardRef } from "react"
+
+interface CardContentProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string
+}
+
+export const CardContent = forwardRef<HTMLDivElement, CardContentProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
+))
